(Implement) select monster and set name
