+++ src/components/monster/MonsterThumbnail.js
