+++ src/recoil/states/monster.js
@@ -1,5 +1,6 @@
 import { atom, selector } from 'recoil';
 import { mainApis, monsterApis } from '../../api';
+import { OK } from '../../constants/statusCode';
 
 const asyncDefaultMonster = selector({
   key: 'asyncDefaultMonster',
@@ -21,10 +22,13 @@ export const monsterState = atom({
 // @jaekyung : character selection page
 const initiateMonsterSelector = selector({
   key: 'initiateMonster',
-  get: async () => {
+  get: async ({ get }) => {
     try {
-      const { data } = await monsterApis.loadStartMonster();
-      return data.monsters;
+      const { data, status } = await monsterApis.loadStartMonster();
+      console.log('data', data);
+      if (status === OK) {
+        return data.monsters;
+      }
     } catch (error) {
       console.error(error);
       throw error;
@@ -33,11 +37,21 @@ const initiateMonsterSelector = selector({
 });
 
 export const babyMonsterState = atom({
-  key: 'babyMonster',
+  key: 'babyMonsterState',
   default: initiateMonsterSelector,
 });
 
-export const selectedMonster = atom({
-  key: 'levelOneMonster',
+export const selectedMonsterState = atom({
+  key: 'selectedMonsterState',
   default: null,
+});
+
+export const getSelectedMonster = selector({
+  key: 'getSelectedMonster',
+  get: ({ get }) => {
+    const selectedMonster = get(selectedMonsterState);
+    console.log('selectedMonster', selectedMonster);
+
+    return selectedMonster;
+  },
 });
